requirements.txt
PyMuPDF
numpy
scikit-learn
tensorflow>=2.0.0
tensorflow-hub
openai
gradio
langchain-serve>=0.0.19

fastapi
tensorflow_hub
fitz
PyMuPDF
requests
urllib3
gql

discord

discord.py==2.1.*
python-dotenv==0.21.*
PyYAML==6.0
dacite==1.6.*